Add files via upload
The combined CSV file houses the initial training and test dataset found on https://archive.ics.uci.edu/ml/datasets/Bank+Marketing. The resampled_Data is a result of the process of handling the class imbalance within the outcome of the previously combined dataset. Both without label encoding and with label encoding.